Improve logging: MDC (#620)
Before:
MDC didn't allow to add more than one value in the same scope and all MDC_PUT were scoped

Now:
- MDC_PUT is not scoped and put value in MDC until removed;
- Introduced SCOPED_MDC* macros for scoped MDC functionality;
- BFT logger changed to mimic the log4cplus behaviour and store MDC in
TLS context object;
- Added thread name to MDC;

+++ threshsign/src/AutoBuf.h
@@ -21,7 +21,7 @@ class AutoBuf {
   int len;
 
  public:
-  AutoBuf(int len) : buf(new T[len]), len(len) {}
+  AutoBuf(int len) : buf(new T[static_cast<size_t>(len)]), len(len) {}
 
   AutoBuf(const AutoBuf<T>& ab) : AutoBuf(ab.len) { std::copy(ab.buf, ab.buf + ab.len, buf); }
